+++ src/common/js/functions.js
@@ -2,23 +2,12 @@
  * pwix:tabular/src/common/js/functions.js
  */
 
-import { check, Match } from 'meteor/check';
+import _ from 'lodash';
 
-/**
- * @param {String} name the tabular name
- * @returns {Array|undefined} the names to be visible on the named tabular
- */
-Tabular.getSettingsColumns = function( tabularName ){
-    check( tabularName, Match.NonEmptyString );
-    let names;
-    if( Meteor.isClient ){
-        const str = Tabular._store.get( COOKIE_COLUMNS, tabularName );
-        if( str ){
-            names = str.split( /,/ );
-        }
-    }
-    return names;
-}
+import { check, Match } from 'meteor/check';
+import { Logger } from 'meteor/pwix:logger';
+
+const logger = Logger.get();
 
 /**
  * @param {Array} columns the original columns as returned by fieldset.toTabular() method